+++ packages/bundle/testBundles/mainFile/main.js
@@ -1,0 +1,10 @@
+export default {
+  extends: {
+    extends: `${__dirname}/extension.js`,
+    buz: {}
+  },
+  dependencies: {
+    bar: {},
+    run: () => jest.fn()
+  }
+}