+++ packages/httpServer/dependencies/index.js
@@ -2,23 +2,26 @@ import joi from '@hapi/joi'
 import Koa from 'koa'
 import Router from 'koa-router'
 import commander from 'commander'
+import pinoHttp from 'pino-http'
+
+import args from './args'
 import createSwaggerDefinition from './swagger/createSwaggerDefinition'
 import errors from './errors'
 import errorMap from './errorMap'
 import events from './events'
 import gracefulShutdown from './gracefulShutdown'
 import health from './health'
+import http from './http'
+import loadRoute from './loadRoute'
 import logger from './logger'
 import middleware from './middleware'
 import requestValidator from './requestValidator'
 import swagger from './swagger'
 import serveStatic from './serveStatic'
-import http from './http'
 import serializers from './serializers'
-import loadRoute from './loadRoute'
 import run from './run'
 import server from './server'
-import args from './args'
+
 export default {
   args,
   commander: () => commander,
@@ -34,6 +37,7 @@ export default {
   logger,
   loadRoute,
   middleware,
+  pinoHttp: () => pinoHttp,
   requestValidator,
   Router: () => Router,
   run,