+++ packages/cli/CHANGELOG.md
@@ -1,4 +1,7 @@
 # CHANGELOG
 
+v0.6.0
+- [CHANGE] Use Bundle and new composable module pattern
+
 v0.5.0
 - [UPDATE] Move package to new namespace